Convert links to array
Convert links to an array and do a map so we don't have to repeat ourself
