+++ src/components/Header/NavbarContent.component.jsx
@@ -1,8 +1,11 @@
 import React from "react"
 import { Link } from "gatsby"
+import { v4 as uuidv4 } from "uuid"
 
 import AppLogo from "./Applogo.component"
 import Hamburger from "./Hamburger.component"
+
+import LINKS from "../../constants/LINKS"
 
 export default function NavbarContent() {
   return (
@@ -12,42 +15,17 @@ export default function NavbarContent() {
         <div className="flex content-center justify-between md:w-1/2 md:justify-end">
           <Hamburger />
           <ul className="items-center justify-between flex-1 hidden list-reset md:flex lg:flex xl:flex">
-            <li className="mr-3">
-              <Link
-                className="inline-block m-4 text-xl text-white hover:underline"
-                activeClassName="underline"
-                to="/"
-              >
-                Hjem
-              </Link>
-            </li>
-            <li className="mr-3">
-              <Link
-                className="inline-block m-4 text-xl text-white hover:underline"
-                activeClassName="underline"
-                to="/portefølje"
-              >
-                Portefølje
-              </Link>
-            </li>
-            <li className="mr-3">
-              <Link
-                className="inline-block m-4 text-xl text-white hover:underline"
-                activeClassName="underline"
-                to="/cv"
-              >
-                CV
-              </Link>
-            </li>
-            <li className="mr-3">
-              <Link
-                className="inline-block m-4 text-xl text-white hover:underline"
-                activeClassName="underline"
-                to="/kontakt"
-              >
-                Kontakt
-              </Link>
-            </li>
+            {LINKS.map((link) => (
+              <li key={uuidv4()} className="mr-3">
+                <Link
+                  className="inline-block m-4 text-xl text-white hover:underline"
+                  activeClassName="underline"
+                  to={link.url}
+                >
+                  {link.text}
+                </Link>
+              </li>
+            ))}
           </ul>
         </div>
       </div>
