+++ src/constants/LINKS.js
@@ -1,0 +1,20 @@
+export const LINKS = [
+  {
+    url: "/",
+    text: "Hjem",
+  },
+  {
+    url: "/Portefølje",
+    text: "Portefølje",
+  },
+  {
+    url: "/CV",
+    text: "CV",
+  },
+  {
+    url: "/Kontakt",
+    text: "Kontakt",
+  },
+]
+
+export default LINKS
