--- src/pages/design1.js
@@ -1,28 +1,0 @@
-/*
-
-Inspirational resources: 
-
-https://www.dcode.no/
-https://www.coretrek.no/
-
-*/
-
-import React from "react"
-
-import Navbar from "../components/Header/Navbar.component"
-import NavbarContent from "../components/Header/NavbarContent.component"
-
-import MainContent from "../components/Design2/MainContent.component"
-
-function IndexPage() {
-  return (
-    <>
-      <Navbar>
-        <NavbarContent />
-      </Navbar>
-      <MainContent />
-    </>
-  )
-}
-
-export default IndexPage
