+++ src/pages/index.js
@@ -12,7 +12,7 @@ import React from "react"
 import Navbar from "../components/Header/Navbar.component"
 import NavbarContent from "../components/Header/NavbarContent.component"
 
-import MainContent from "../components/Design3/MainContent.component"
+import MainContent from "../components/Index/MainContent.component"
 
 function IndexPage() {
   return (
